Extract ownership reconciliation from orchestrator
